+++ docs/changes.rst
@@ -53,6 +53,8 @@ greatly improved MPD client support.
   - Rename :mod:`mopidy.backends.gstreamer` to :mod:`mopidy.backends.local`.
   - Remove :mod:`mopidy.backends.despotify`, as Despotify is little maintained
     and the Libspotify backend is working much better.
+  - Rename ``mopidy.frontends.mpd.{serializer => translator}`` to match naming
+    in backends.
 
 - Backend API: